Create Boats to Save People.java
You are given an array of people where people[i] is the weight of the ith person and an infinite number of boats where each boat can carry a maximum weight of limit. Each boat carries at most two people at the same time, provided the sum of the weight of those people is at the most limit.

Return the minimum number of boats to carry every given person.